+++ src/browser_policy.rs
@@ -215,7 +215,6 @@ impl BrowserPolicy {
                     };
                 }
             }
-            _ => {}
         }
         JsValidation::Allowed
     }